LPS-183879 Moves logic from portal kernel to layout set prototype modules
