+++ portal-kernel/src/com/liferay/sites/kernel/util/packageinfo
@@ -1,1 +1,1 @@
-version 1.5.0
+version 1.4.0
